+++ yolo_bit/include/day_night_mode.h
@@ -1,0 +1,8 @@
+#ifndef DAY_NIGHT_MODE_H
+#define DAY_NIGHT_MODE_H
+
+#include "global.h"
+
+void day_night_task(void *pvParameters);
+
+#endif // DAY_NIGHT_MODE_H