+++ yolo_bit/include/secured_door.h
@@ -1,0 +1,8 @@
+#ifndef SECURED_DOOR_H
+#define SECURED_DOOR_H
+
+#include "global.h"
+
+void secured_door_task(void *pvParameters);
+
+#endif // SECURED_DOOR_H