+++ yolo_bit/docs/CHANGELOG.md
@@ -3,5 +3,21 @@
 ## v0.0.1 - 19/03/2026
 
 - Initial project setup and documentation.
-- **Added Day/Night Power Saving Mode (Automatic):** Configured NeoPixel LEDs and Light Sensor for automatic lighting control based on ambient darkness.
-- **Added Temperature & Humidity Monitoring:** Integrated DHT20 sensor to track environmental variables and periodically log values (Temp, Humid, and Light Level) to the Serial Monitor every 5 seconds.
+- **Modularized Codebase (RTOS):** Disassembled `main.cpp` logic into discrete FreeRTOS tasks.
+- **Added Day/Night Task:** Implemented in `day_night_mode.cpp/h`.
+- **Added Temp/Humi Task:** Implemented in `temp_humi_monitor.cpp/h`.
+- **Added Secured Door Task:** Scaffolding implemented in `secured_door.cpp/h`.
+- **Added AppContext:** Utilizing Semaphores for IPC in `global.h` & `config.h`.
+- **Change include** paths in `global.h` from:
+
+```cpp
+//this cause a not-found error
+#include <FreeRTOS.h>
+#include <semphr.h>
+
+//to
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+#include "freertos/semphr.h"
+//this fix everything
+```